+++ source/includes/authentication/aws-eks-connection-string.py
@@ -1,0 +1,15 @@
+import os
+from pymongo import MongoClient
+from pymongo.auth_oidc import OIDCCallback, OIDCCallbackContext, OIDCCallbackResult
+
+# define callback, properties, URI, and MongoClient
+class MyCallback(OIDCCallback):
+   def fetch(self, context: OIDCCallbackContext) -> OIDCCallbackResult:
+       with open(os.environ["AWS_WEB_IDENTITY_TOKEN_FILE"]) as fid:
+           token = fid.read()
+       return OIDCCallbackResult(access_token=token)
+properties = {"OIDC_CALLBACK": MyCallback()}
+uri = ("mongodb://<hostname>:<port>/?"
+       "&authMechanism=MONGODB-OIDC"
+       "&authMechanismProperties=properties")
+client = MongoClient(uri)
