--- source/includes/authentication/azure-imds.py
@@ -1,24 +1,0 @@
-# start mongoclient
-from pymongo import MongoClient
-
-# define properties and MongoClient
-properties = {"ENVIRONMENT": "azure", "TOKEN_RESOURCE": "<audience>"}
-client = MongoClient(
-    "mongodb://<hostname>:<port>",
-    username="<Azure identity client ID>",
-    authMechanism="MONGODB-OIDC",
-    authMechanismProperties=properties,
-)
-# end mongoclient
-
-# start connection string
-from pymongo import MongoClient
-
-# define properties, URI, and MongoClient
-properties = {"ENVIRONMENT": "azure", "TOKEN_RESOURCE": "<audience>"}
-uri = ("mongodb://<hostname>:<port>/?"
-       "username=<Azure identity client ID>"
-       "&authMechanism=MONGODB-OIDC"
-       "&authMechanismProperties=properties")
-client = MongoClient(uri)
-# end connection string
