--- source/includes/authentication/gcp-gke.py
@@ -1,34 +1,0 @@
-# start mongoclient
-from pymongo import MongoClient
-from pymongo.auth_oidc import OIDCCallback, OIDCCallbackContext, OIDCCallbackResult
-         
-# define callback, properties, and MongoClient
-class MyCallback(OIDCCallback):
-    def fetch(self, context: OIDCCallbackContext) -> OIDCCallbackResult:
-        with open("/var/run/secrets/kubernetes.io/serviceaccount/token") as fid:
-            token = fid.read()
-        return OIDCCallbackResult(access_token=token)
-properties = {"OIDC_CALLBACK": MyCallback()}
-client = MongoClient(
-   "mongodb://<hostname>:<port>",
-   authMechanism="MONGODB-OIDC",
-   authMechanismProperties=properties,
-)
-# end mongoclient
-
-# start connection string
-from pymongo import MongoClient
-from pymongo.auth_oidc import OIDCCallback, OIDCCallbackContext, OIDCCallbackResult
-
-# define callback, properties, URI, and MongoClient
-class MyCallback(OIDCCallback):
-    def fetch(self, context: OIDCCallbackContext) -> OIDCCallbackResult:
-        with open("/var/run/secrets/kubernetes.io/serviceaccount/token") as fid:
-            token = fid.read()
-        return OIDCCallbackResult(access_token=token)
-properties = {"OIDC_CALLBACK": MyCallback()}
-uri = ("mongodb://<hostname>:<port>/?"
-       "&authMechanism=MONGODB-OIDC"
-       "&authMechanismProperties=properties")
-client = MongoClient(uri)
-# end connection string
